a shitload of things
first of all, the code needs serious refactoring. That's the top priority once I'm done with the naming system

it still needs layers

maybe Vuforia in DEVICE mode is the way to go to solve the problem of text falling apart, but it requires a bit figuring out about how to attach text to trackable without it falling apart in another way.

AR Periodic Table has started